pomni/backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
mysql==0.0.3
mysqlclient==2.2.7
sqlparse==0.5.3
tzdata==2025.2
